Speed up InterpreterSelectWithUnionQuery compilation
