+++ src/AggregateFunctions/parseAggregateFunctionParameters.h
@@ -1,7 +1,7 @@
 #pragma once
 
+#include <Core/Field.h>
 #include <Interpreters/Context_fwd.h>
-#include <Parsers/ASTExpressionList.h>
 #include <Parsers/IAST_fwd.h>
