+++ src/Common/OpenTelemetryTraceContext.h
@@ -1,5 +1,8 @@
 #pragma once
 
+#include <base/types.h>
+#include <base/UUID.h>
+
 namespace DB
 {
