+++ src/Core/ColumnWithTypeAndName.cpp
@@ -1,10 +1,18 @@
+#include <Columns/IColumn.h>
 #include <Core/ColumnsWithTypeAndName.h>
-#include <IO/WriteBufferFromString.h>
+#include <DataTypes/IDataType.h>
 #include <IO/Operators.h>
+#include <IO/WriteBufferFromString.h>
 
 
 namespace DB
 {
+
+/// Uses type->createColumn() to create column
+ColumnWithTypeAndName::ColumnWithTypeAndName(const DataTypePtr & type_, const String & name_)
+    : column(type_->createColumn()), type(type_), name(name_)
+{
+}
 
 ColumnWithTypeAndName ColumnWithTypeAndName::cloneEmpty() const
 {
