+++ src/Parsers/Access/ASTCreateRoleQuery.h
@@ -6,6 +6,7 @@
 
 namespace DB
 {
+using Strings = std::vector<String>;
 class ASTSettingsProfileElements;
