+++ src/Parsers/Access/ASTUserNameWithHost.h
@@ -6,6 +6,8 @@
 namespace DB
 {
 
+using Strings = std::vector<String>;
+
 /** Represents a user name.
   * It can be a simple string or identifier or something like `name@host`.
   * In the last case `host` specifies the hosts user is allowed to connect from.
