+++ src/Processors/Transforms/DistinctSortedTransform.h
@@ -1,8 +1,9 @@
 #pragma once
 
-#include <Processors/ISimpleTransform.h>
-#include <Interpreters/SetVariants.h>
 #include <Core/SortDescription.h>
+#include <Interpreters/SetVariants.h>
+#include <Processors/ISimpleTransform.h>
+#include <QueryPipeline/SizeLimits.h>
 
 
 namespace DB
