+++ src/Storages/FileLog/FileLogDirectoryWatcher.h
@@ -6,6 +6,7 @@
 
 #include <memory>
 #include <mutex>
+#include <unordered_map>
 
 namespace DB
 {
